backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# Config
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Auth
python-jose[cryptography]>=3.3.0

# Security
cryptography>=42.0.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9

# HTTP client (used by multi-agent LLM calls)
httpx>=0.27.0

# Background tasks
arq>=0.25.0
prometheus-fastapi-instrumentator>=6.1.0
prometheus-client>=0.20.0
